chore: one last overhaul
