+++ .github/workflows/pre-release.yml
@@ -59,37 +59,8 @@ jobs:
       - run: npx nx run-many -t build --parallel=10
         env:
           NX_CLOUD_ACCESS_TOKEN: ${{ secrets.NX_CLOUD_ACCESS_TOKEN }}
-      - run: git push --follow-tags --dry-run
       - run: npx nx run-many -t can-publish --verbose --parallel=40
       - run: git push --follow-tags
       - run: npx nx release publish
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NODE_AUTH_TOKEN }}
-      # - name: Check if everything builds with nx release cache
-      #   run: npx nx affected:build --parallel=3
-      #   env:
-      #     NX_CLOUD_ACCESS_TOKEN: ${{ secrets.NX_CLOUD_RELEASE_TOKEN }}
-      #   run: >-
-      #     npx nx run dashboard:versionCI
-      #     --releaseAs=$(
-      #     git cherry -v main | grep -Eq "chore: release version.*${{ github.event.inputs.releaseAs }}.*"
-      #     && echo "prerelease"
-      #     || echo ${{ github.event.inputs.releaseAs }})
-      #     --preid=${{ github.event.inputs.preid }}
-      #     --skip-nx-cache
-      # env:
-      #   NX_CLOUD_ACCESS_TOKEN: ${{ secrets.NX_CLOUD_RELEASE_TOKEN }}
-
-      # - name: Check for created tags to see if a new version has been made
-      #   run: git tag -l --contains $(git rev-parse HEAD) | cat | grep . && echo "Found newly created tag" || { echo "No changes found"; exit 1; };
-
-      # - name: Push changes
-      #   run: git push origin HEAD
-      # - name: Push tags
-      #   run: git push --tags
-
-      # - name: Publish changes
-      #   run: npx nx affected --target=publishCI --parallel 1 --nx-bail
-      #   env:
-      #     NODE_AUTH_TOKEN: ${{ secrets.NODE_AUTH_TOKEN }}
-      #     NX_CLOUD_ACCESS_TOKEN: ${{ secrets.NX_CLOUD_RELEASE_TOKEN }}
