+++ .github/workflows/release.yml
@@ -52,26 +52,8 @@ jobs:
       - run: npx nx run-many -t build --parallel=10
         env:
           NX_CLOUD_ACCESS_TOKEN: ${{ secrets.NX_CLOUD_ACCESS_TOKEN }}
-      - run: git push --follow-tags --dry-run
       - run: npx nx run-many -t can-publish --parallel=40
+      - run: git push --follow-tags
       - run: npx nx release publish
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NODE_AUTH_TOKEN }}
-      - run: git push --follow-tags
-      # - name: Version
-      #   run: npx nx run dashboard:versionCI --skip-nx-cache
-      # - name: Check for created tags to see if a new version has been made
-      #   run: git tag -l --contains $(git rev-parse HEAD) | cat | grep . && echo "Found newly created tag" || { echo "No changes found"; exit 1; };
-      # - name: Publish changes
-      #   run: npx nx affected --target=publishCI --parallel 1 --nx-bail
-      #   env:
-      #     NODE_AUTH_TOKEN: ${{ secrets.NODE_AUTH_TOKEN }}
-      #     NX_CLOUD_ACCESS_TOKEN: ${{ secrets.NX_CLOUD_RELEASE_TOKEN }}
-      # - name: Push tags
-      #   run: git push --tags
-      # - name: Push changes
-      #   run: git push origin HEAD
-      # - name: Tag last-release
-      #   run: |
-      #     git tag -f last-release
-      #     git push origin last-release --force
