+++ libs/browser/README.md
@@ -1,11 +1,3 @@
-# browser
+# @dynamic-msw/browser
 
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build browser` to build the library.
-
-## Running unit tests
-
-Run `nx test browser` to execute the unit tests via [Vitest](https://vitest.dev/).
+Documentation can be found here: [dynamic-msw](https://github.com/dynamicmsw/dynamic-msw/tree/main?tab=readme-ov-file#dynamic-mock-service-worker)
