+++ libs/core/README.md
@@ -1,12 +1,3 @@
-# core
+# @dynamic-msw/core
 
-This library was generated with [Nx](https://nx.dev).
-s
-
-## Building
-
-Run `nx build core` to build the library.
-
-## Running unit tests
-
-Run `nx test core` to execute the unit tests via [Vitest](https://vitest.dev/).
+Documentation can be found here: [dynamic-msw](https://github.com/dynamicmsw/dynamic-msw/tree/main?tab=readme-ov-file#dynamic-mock-service-worker)
