+++ libs/core/tsconfig.lib.json
@@ -6,5 +6,5 @@
     "types": ["node", "vite/client"]
   },
   "include": ["src/**/*.ts"],
-  "exclude": ["vite.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"]
+  "exclude": ["vite.config.mts", "src/**/*.spec.ts", "src/**/*.test.ts"]
 }
