+++ libs/dashboard-button/vite.config.mts
@@ -2,18 +2,29 @@
 import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
 import * as path from 'path';
+import fs from 'fs';
 import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
+import { viteStaticCopy } from 'vite-plugin-static-copy';
+
+process.env.VITE_DASHBOARD_HTML = getDashboardHTML();
 
 export default defineConfig({
   root: __dirname,
-  cacheDir: '../../node_modules/.vite/core',
-
+  cacheDir: '../../node_modules/.vite/dashboard-button',
   plugins: [
     nxViteTsPaths(),
     dts({
       entryRoot: 'src',
       tsConfigFilePath: path.join(__dirname, 'tsconfig.lib.json'),
       skipDiagnostics: true,
+    }),
+    viteStaticCopy({
+      targets: [
+        {
+          src: 'README.md',
+          dest: '',
+        },
+      ],
     }),
   ],
 
@@ -25,7 +36,7 @@ export default defineConfig({
   // Configuration for building your library.
   // See: https://vitejs.dev/guide/build.html#library-mode
   build: {
-    outDir: '../../dist/core',
+    outDir: '../../dist/dashboard-button',
     reportCompressedSize: true,
     commonjsOptions: {
       transformMixedEsModules: true,
@@ -33,7 +44,7 @@ export default defineConfig({
     lib: {
       // Could also be a dictionary or array of multiple entry points.
       entry: 'src/index.ts',
-      name: 'core',
+      name: 'dashboard-button',
       fileName: 'index',
       // Change this to the formats you want to support.
       // Don't forget to update your package.json as well.
@@ -45,3 +56,14 @@ export default defineConfig({
     },
   },
 });
+
+function getDashboardHTML(): string {
+  try {
+    return fs.readFileSync(
+      path.join(__dirname, '../../dist/dashboard/index.html'),
+      'utf-8'
+    );
+  } catch {
+    return '<div>Fallback: Dashboard build output is missing</div>';
+  }
+}
