+++ libs/dashboard-core/README.md
@@ -1,11 +1,3 @@
-# dashboard-core
+# @dynamic-msw/dashboard-core
 
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build dashboard-core` to build the library.
-
-## Running unit tests
-
-Run `nx test dashboard-core` to execute the unit tests via [Vitest](https://vitest.dev/).
+Documentation can be found here: [dynamic-msw](https://github.com/dynamicmsw/dynamic-msw/tree/main?tab=readme-ov-file#dynamic-mock-service-worker)
