+++ libs/dashboard/README.md
@@ -1,3 +1,3 @@
-# @dynamic-msw/setup
+# @dynamic-msw/dashboard
 
 Documentation can be found here: [dynamic-msw](https://github.com/dynamicmsw/dynamic-msw/tree/main?tab=readme-ov-file#dynamic-mock-service-worker)
