+++ libs/dashboard/tsconfig.spec.json
@@ -13,7 +13,7 @@
     ]
   },
   "include": [
-    "vite.config.ts",
+    "vite.config.mts",
     "vitest.config.ts",
     "src/**/*.test.ts",
     "src/**/*.spec.ts",
