--- libs/dynamic-msw/README.md
@@ -1,11 +1,0 @@
-# dynamic-msw
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build dynamic-msw` to build the library.
-
-## Running unit tests
-
-Run `nx test dynamic-msw` to execute the unit tests via [Vitest](https://vitest.dev/).
