+++ libs/dynamic-msw/tsconfig.spec.json
@@ -11,7 +11,7 @@
     ]
   },
   "include": [
-    "vite.config.ts",
+    "vite.config.mts",
     "vitest.config.ts",
     "src/**/*.test.ts",
     "src/**/*.spec.ts",
