+++ libs/dynamic-msw/vite.config.mts
@@ -3,10 +3,11 @@ import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
 import * as path from 'path';
 import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
+import { viteStaticCopy } from 'vite-plugin-static-copy';
 
 export default defineConfig({
   root: __dirname,
-  cacheDir: '../../node_modules/.vite/node',
+  cacheDir: '../../node_modules/.vite/dynamic-msw',
 
   plugins: [
     nxViteTsPaths(),
@@ -14,6 +15,14 @@ export default defineConfig({
       entryRoot: 'src',
       tsConfigFilePath: path.join(__dirname, 'tsconfig.lib.json'),
       skipDiagnostics: true,
+    }),
+    viteStaticCopy({
+      targets: [
+        {
+          src: '../../README.md',
+          dest: '',
+        },
+      ],
     }),
   ],
 
@@ -25,7 +34,7 @@ export default defineConfig({
   // Configuration for building your library.
   // See: https://vitejs.dev/guide/build.html#library-mode
   build: {
-    outDir: '../../dist/node',
+    outDir: '../../dist/dynamic-msw',
     reportCompressedSize: true,
     commonjsOptions: {
       transformMixedEsModules: true,
@@ -33,7 +42,7 @@ export default defineConfig({
     lib: {
       // Could also be a dictionary or array of multiple entry points.
       entry: 'src/index.ts',
-      name: 'node',
+      name: 'dynamic-msw',
       fileName: 'index',
       // Change this to the formats you want to support.
       // Don't forget to update your package.json as well.
@@ -41,22 +50,7 @@ export default defineConfig({
     },
     rollupOptions: {
       // External packages that should not be bundled into your library.
-      external: [/node_modules/, 'msw/node'],
-    },
-  },
-
-  test: {
-    globals: true,
-    cache: {
-      dir: '../../node_modules/.vitest',
-    },
-    environment: 'node',
-    include: ['src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
-
-    reporters: ['default'],
-    coverage: {
-      reportsDirectory: '../../coverage/node',
-      provider: 'v8',
+      external: [/node_modules/],
     },
   },
 });
