+++ libs/setup/vite.config.mts
@@ -2,20 +2,27 @@
 import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
 import * as path from 'path';
-import fs from 'fs';
 import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
-
-process.env.VITE_DASHBOARD_HTML = getDashboardHTML();
+import { viteStaticCopy } from 'vite-plugin-static-copy';
 
 export default defineConfig({
   root: __dirname,
-  cacheDir: '../../node_modules/.vite/dashboard-button',
+  cacheDir: '../../node_modules/.vite/setup',
+
   plugins: [
     nxViteTsPaths(),
     dts({
       entryRoot: 'src',
       tsConfigFilePath: path.join(__dirname, 'tsconfig.lib.json'),
       skipDiagnostics: true,
+    }),
+    viteStaticCopy({
+      targets: [
+        {
+          src: 'README.md',
+          dest: '',
+        },
+      ],
     }),
   ],
 
@@ -27,7 +34,7 @@ export default defineConfig({
   // Configuration for building your library.
   // See: https://vitejs.dev/guide/build.html#library-mode
   build: {
-    outDir: '../../dist/dashboard-button',
+    outDir: '../../dist/setup',
     reportCompressedSize: true,
     commonjsOptions: {
       transformMixedEsModules: true,
@@ -35,7 +42,7 @@ export default defineConfig({
     lib: {
       // Could also be a dictionary or array of multiple entry points.
       entry: 'src/index.ts',
-      name: 'dashboard-button',
+      name: 'setup',
       fileName: 'index',
       // Change this to the formats you want to support.
       // Don't forget to update your package.json as well.
@@ -47,14 +54,3 @@ export default defineConfig({
     },
   },
 });
-
-function getDashboardHTML(): string {
-  try {
-    return fs.readFileSync(
-      path.join(__dirname, '../../dist/dashboard/index.html'),
-      'utf-8'
-    );
-  } catch {
-    return '<div>Fallback: Dashboard build output is missing</div>';
-  }
-}
