+++ package.json
@@ -82,6 +82,7 @@
     "vite": "~5.0.0",
     "vite-plugin-dts": "~2.3.0",
     "vite-plugin-singlefile": "^2.0.0",
+    "vite-plugin-static-copy": "^1.0.1",
     "vitest": "^1.0.4"
   },
   "nx": {
